fix: Category 수정
1. Page를 여러 개 받아온 상태에서 Tag를 클릭하면 기존 데이터로 덮어지는 버그 수정
2. 특정 기능들 함수로 분할
3. IntersectionObserver Custom Hook 분할
4. 존재하지 않는 데이터 처리
